Move page_viewport() into gfx::viewport and reduce it to orchestration
