+++ crates/browser/src/lib.rs
@@ -1,6 +1,4 @@
 pub mod form_controls;
-pub mod input_store;
-pub mod interactions;
 pub mod page;
 pub mod resources;
 pub mod shell;
