+++ crates/gfx/Cargo.toml
@@ -12,3 +12,5 @@ pollster = { workspace = true }
 css = { path = "../css" }
 layout = { path = "../layout" }
 core_types = { path = "../core_types" }
+html = { path = "../html" }
+url = { workspace = true }
