+++ crates/gfx/src/input/mod.rs
@@ -1,0 +1,11 @@
+mod interaction;
+mod route;
+mod store;
+
+pub use interaction::{
+    ActiveTarget, InputDragState, InteractionState, TextareaCachedLine, TextareaCachedTextFragment,
+    TextareaLayoutCache,
+};
+pub use route::FormControlHandler;
+pub(crate) use route::{FrameInputCtx, route_frame_input};
+pub use store::{InputValueStore, SelectionRange};
