+++ crates/gfx/src/input/store.rs
